+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All versions prior to 0.0.9 are untracked.
 
 ## [Unreleased]
 
+## [2.3.1] - 2022-05-24
+
 ### Fixed
 
 * CLI: A bug causing the terminal's cursor to disappear on some
@@ -223,6 +225,7 @@ All versions prior to 0.0.9 are untracked.
 
 <!-- Release URLs -->
 [Unreleased]: https://github.com/trailofbits/pip-audit/compare/v2.0.0...HEAD
+[2.3.1]: https://github.com/trailofbits/pip-audit/compare/v2.3.0...v2.3.1
 [2.3.0]: https://github.com/trailofbits/pip-audit/compare/v2.2.1...v2.3.0
 [2.2.1]: https://github.com/trailofbits/pip-audit/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/trailofbits/pip-audit/compare/v2.1.1...v2.2.0